added delete functionality --- [last change in project]

+++ routes/ideas.js
@@ -48,16 +48,24 @@ ideaRouter.post("/", async (req, res) => {
 ideaRouter.put("/:id", async (req, res) => {
 
     try {
-        const updatedIdea = await Idea.findByIdAndUpdate(req.params.id,
-            {
-                $set: {
-                    text: req.body.text,
-                    tag: req.body.tag
-                }
-            }, { new: true });
+
+        const idea = await Idea.findById(req.params.id);
+
+        // Match the usernames
+        if (idea.username === req.body.username) {
+            const updatedIdea = await Idea.findByIdAndUpdate(req.params.id,
+                {
+                    $set: {
+                        text: req.body.text,
+                        tag: req.body.tag
+                    }
+                }, { new: true });
 
 
-        res.json({ status: true, data: updatedIdea });
+            return res.json({ status: true, data: updatedIdea });
+        }
+        res.status(403).json({ success: false, error: "You are not authorized to update this resource" });
+
     } catch (err) {
         res.status(500).json({ success: false, error: "Something went wrong" });
     }
@@ -68,8 +76,18 @@ ideaRouter.put("/:id", async (req, res) => {
 ideaRouter.delete("/:id", async (req, res) => {
 
     try {
-        const deletedIdea = await Idea.findByIdAndDelete(req.params.id);
-        res.json({ status: true, data: deletedIdea });
+
+        const idea = await Idea.findById(req.params.id);
+
+        // Match the usernames
+        if (idea.username === req.body.username) {
+            await Idea.findByIdAndDelete(req.params.id);
+            return res.json({ status: true, data: {} });
+        }
+
+        // Usernames do not match
+        res.status(403).json({ success: false, error: "You are not authorized to delete this resource " });
+
     } catch (err) {
         res.status(500).json({ success: false, error: "Something went wrong" });
     }
